Atskirtas vilkiku valdymo modulis
